Implement local storage for resumable metrics
Store collected SP information on disk such that on service restart data
and stats is reloaded. This would avoid long wait in order of hours to
wait for chain walk before metrics get populated.

Silently fall back on populating the metrics if loading the stored info
for a SP fails.